% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_souls.R
\name{process_souls}
\alias{process_souls}
\title{Processing of raw LCMS data using SOULS (Segmentation of untargeted LCMS spectra)}
\usage{
process_souls(
  path_mzMLs,
  path_Ref,
  index_ref = 1,
  path_csv,
  num_workers,
  RT_range,
  size_RT_segs,
  mz_range,
  size_mz_segs
)
}
\arguments{
\item{path_mzMLs}{Path to the folder where the mzML files of the samples are located.}

\item{path_Ref}{Path to the folder containing the reference sample.}

\item{index_ref}{Index of the reference sample in the folder; Default: 1}

\item{path_csv}{Path to the csv file containing the pheno-data such as name, instrument, origin, variety...}

\item{num_workers}{Number of cores used for parallel processing. Recommendation: 6/96 or 2/8 Cores; Default: 2}

\item{RT_range}{Retention time range of the LCMS spectra; Default: c(300, 360)}

\item{size_RT_segs}{Size of the segments at retention time level; Default: 10}

\item{mz_range}{m/z range of the LCMS spectra; Default: c(250, 750)}

\item{size_mz_segs}{Size of the segments at m/z level; Default: 5}
}
\value{
Matrix showing summed intensities in each segment.
}
\description{
Processing of raw LCMS data using SOULS (Segmentation of untargeted LCMS spectra)
}
\examples{
seg.result <- process_souls(path_mzMLs = "path/to/folder/Samples", path_Ref = "path/to/folder/Ref", path_csv = "path/to/folder/phenodata")
}
