% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/souls.R
\name{souls}
\alias{souls}
\title{Segmentation of untargeted LCMS data}
\usage{
souls(object, num_workers, RT_range, size_RT_segs, mz_range, size_mz_segs)
}
\arguments{
\item{object}{XCMSnExp object from xcms workflow (https://bioconductor.org/packages/release/bioc/vignettes/xcms/inst/doc/xcms.html) after retention time alignment (the SOULS step replaces the correspondence step in the xcms workflow).}

\item{RT_range}{Retention time range of the spectra to be processed.}

\item{size_RT_segs}{Size of the segments in retention time dimension (RT_range should be divisible by size_RT_segs).}

\item{mz_range}{Mass range of the spectra to be processed.}

\item{size_mz_segs}{Size of the segments in mass dimension (mz_range should be divisible by size_mz_segs)}
}
\value{
Matrix showing summed intensities in each segment.
}
\description{
Segmentation of untargeted LCMS data
}
\examples{
seg.result <- souls(object = xdata_adj, RT_range = c(300, 360), size_RT_segs = 10, mz_range = mz_range = c(250, 750), size_mz_segs = 5)
}
